Moved verifier related tests to the verifier folder
